Avoid Torch warning/error messages when managing arrays that are already on Torch.
The old `torch.tensor` converter did just the right thing: construct a new
tensor when given lists or NumPy arrays, and simply clone the input if it was
already a PyTorch tensor.
For some reason, Torch has deprecated this behaviour, so it is now necessary
to hard-code the different possibilities.