Generalize Astra detector-curvature special-cases to not crash when the geometry lacks the `det_curvature_radius` attribute.
This attribute is currently only present for the cone-like geometries, with the
consequence that the parallel-beam tests fail for no reason except that the checking
code cannot check this attribute. Effectively, this corresponds to the `None` case
anyway though, so that the problem can be circumvented with the soft-lookup performed
by `getattr` with a default value.

A more principled alternative would be to require an explicit `det_curvature_radius ≡ None`
for the parallel geometries, perhaps with a better OO hierarchy of geometry classes.
But it is doubtful whether this would be worth the effort.