Idea for how a separate `DiscretizedSpace` class could be generated around each backend-specific `TensorSpace`.
The motivation is that it would allow the properties like `available_dtypes` to remain static (not requiring instantiation beforehand. It would also have some other nice effects, in particular there is no need to explicitly carry around a `tensor` and delegate content-related method calls to it, because the class would directly inherit from the concrete tensor class.

It relies on rather aggressive exploitation of Python's dynamic architecture though, and the disadvantages probably outweigh.