Ensure operators return something when called out-of-place.
The operator in the test suite did not return anything from its `_call` method.
This was then patched over in `__call__` with self.range.element(out), which
in NumPy happened to give the original values of the `out` array - but as
far as we can see, this was pure luck.

For reliable behaviour, all operators should actually have a `return` statement.